# Stop copying LogicalPlan and Exprs in `EliminateCrossJoin` (4% faster planning)